pyproject.toml
[project]
name = "businesslake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.0.0",
    "duckdb>=1.1.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "langchain>=0.3.1",
    "langchain-openai>=0.1.23",
    "matplotlib>=3.9.2",
    "nltk>=3.9.1",
    "pandas>=2.2.2",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.1.0",
    "sentencepiece>=0.2.0",
    "sqlglot[rs]>=25.20.1",
    "sqlparse>=0.5.1",
    "tabulate>=0.9.0",
    "torch>=2.4.1",
    "tqdm>=4.66.5",
    "transformers>=4.44.2",
    "protobuf<=5.0.0",
    "langchain-google-genai==2.0.0",
    "langchain-ollama>=0.2.0",
    "faiss-cpu>=1.8.0.post1",
    "langchain-community>=0.3.1",
    "openpyxl>=3.1.5",
    "spacy>=3.8.3",
    "pip>=24.3.1",
]
